\name{SynInt-package}
\alias{SynInt-package}
\alias{SynInt}
\docType{package}
\title{
\packageTitle{SynInt}
}
\description{
\packageDescription{SynInt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SynInt}
\packageIndices{SynInt}
SIMsampler samples from the posterior distribution of the synergistic interaction model.
}
\author{
\packageAuthor{SynInt}

Maintainer: \packageMaintainer{SynInt}
}
\references{
[to be added]
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
\code{\link{SIMsampler}}
}
\examples{
#### To be added ####
}
